p4i integration Integrate 234545 from 4.2 to main: Fix bug where the pen width would be truncated to an int because adjustPenWidth was defined to take an int argument instead of a float. Also, skip adjusting the pen if we have a complex transform OR antialiasing is enabled.
[git-p4: depot-paths = "//depot/qt/main/": change = 234546]

+++ src/gui/painting/qpaintengine_mac_p.h
@@ -143,7 +143,7 @@ public:
     void setStrokePen(const QPen &pen);
     float penOffset();
     QPointF devicePixelSize(CGContextRef context);
-    float adjustPenWidth(int penWidth);
+    float adjustPenWidth(float penWidth);
     inline void setTransform(const QMatrix *matrix=0)
     {
         CGContextConcatCTM(hd, CGAffineTransformInvert(CGContextGetCTM(hd)));
